+++ js/dom-utils.js
@@ -1,0 +1,9 @@
+export function loadingAnimationToggle(display) {
+  const loadingAnimation = document.querySelector('.letter-holder');
+loadingAnimation.style = display;
+}
+
+export function setInputVal(value) {
+  const userSearch = document.querySelector('.search-form__content__input');
+  userSearch.value = value;
+}
