+++ js/index.js
@@ -1,5 +1,5 @@
 import { handleHash } from './router.js';
-import { initHome } from './home.js';
+import { initHome } from './pages/home.js';
 
 document.addEventListener("DOMContentLoaded", function () {
   initHome();
